Se agrego login

+++ src/listaAlumnos/DeleteServlet.java
@@ -2,7 +2,8 @@ package listaAlumnos;
 
 import java.io.IOException;  
 import javax.servlet.ServletException;  
-import javax.servlet.annotation.WebServlet;  
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServlet;  
 import javax.servlet.http.HttpServletRequest;  
 import javax.servlet.http.HttpServletResponse; 
@@ -12,8 +13,17 @@ public class DeleteServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {  
-        int id = Integer.parseInt(request.getParameter("id"));  
-        AlumDAO.delete(id);  
-        response.sendRedirect("ViewServlet");  
+		Cookie ck[]=request.getCookies();  
+	    if(ck!=null){  
+	    	String name=ck[0].getValue();  
+	        if(!name.equals("")||name!=null){ 
+	        	int id = Integer.parseInt(request.getParameter("id"));  
+	            AlumDAO.delete(id);  
+	            response.sendRedirect("ViewServlet"); 
+	        }
+	    }else {
+            request.getRequestDispatcher("login.html").include(request, response);  
+	    }
+         
     }  
 }
